[Android] CoinListScreen enhancement
